Section-2-project/django_auth/api/serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth.password_validation import validate_password
from . import models

User = get_user_model()

class UserSerializer(serializers.ModelSerializer):
    email = serializers.EmailField(required=True)
    password = serializers.CharField(write_only=True, required=True, validators=[validate_password])
    username = serializers.CharField(required=True)

    class Meta:
        model = User
        fields = ('email', 'username', 'password')

    def create(self, validated_data):
        user = User.objects.create_user(
            email=validated_data['email'],
            username=validated_data['username'],
            password=validated_data['password']
        )
        return user
    
class TokenSerializer(serializers.Serializer):
    access = serializers.CharField()
    refresh = serializers.CharField()

    def validate(self, attrs):
        token = RefreshToken.for_user(attrs['user'])
        return {
            'access': str(token.access_token),
            'refresh': str(token.refresh_token),
        }

class UserListSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email', 'date_joined')
        
class BlockUserSerializer(serializers.Serializer):
    email = serializers.EmailField(required= True)
    
class UnblockUserSerializer(serializers.Serializer):
    email = serializers.EmailField(required= True)
    
        
