Section-2-project/django_auth/requirements.txt
Django>=3.0,<4.0
djangorestframework
djangorestframework-simplejwt  # For JWT authentication
dj-rest-auth
django-cors-headers
